[Serving] Refactor to consolidate new request prefill
